feat: basic pokemon lookup
Adds a very basic pokemon lookup command to interact with PokeAPI. This
is NOT meant to be the final implementation and mainly exists for ease
of testing.

Long-term, this won't be tied to the PokeAPI data model and will be
using our own internal data model for Pokemon since we need to filter
returned data to just what's applicable to the generation/version-group
the user has selected.